// ISA-Front/src/app/model/new-fishing-action.model.ts
export class NewFishingAction {
    start: Date;
    durationInDays: number;
    validUntilAndIncluding: Date;
    maxPeople: number;
    price: number;
    address: string;
    city: string;
    country: string;
    longitude: number;
    latitude: number;
    fishingTripReservationTags: string[];
}